Refactor authentication and login flow
Well, I came to find this codebase a bit of a mess, and I needed to
create some clarity about what goes where and how it all works. This is
that.

Previously, the QrCodeScannerWork object had way too many
responsibilities and was passed around awkwardly around the view
hierarchy for no good reason (and also most authentication simply
overlaid the next screen atop the previous one, which meant that the
resources associated with e.g. the QR scanner could never be freed until
the app was re-launched, and I'm not sure whether even that worked for
the kitchen.)

Hence, to clean that up, a new AuthManager ambient object is introduced
into the environment at the topmost level. It globally records the
current authentication state and its associated data (the restaurant and
user objects).

Do note that this only touches authentication (i.e., scanning the
login QR code and selecting the waiter user) and therefore a lot of the
rest of the codebase still relies on fetching documents from Firebase
all the time and referencing them directly by skimming their IDs from
UserDefaults, which is something that _really_ shouldn't be done.

There's some other miscellania touched and/or cleaned up, in most cases
involving whitespace consistency.

A typo fix pass is to be done too.