Rework AuthManager & other future handling + more...
Sorry for the nondescript commit title. I'll do my best to explain what
happened and for what reason in this message.

First of all, fix LaunchScreen to refer to KitchenTabView instead of the
raw KitchenView. Add the debug menu to KitchenTabView.

Add an option to switch to waiter/kitchen from the debug menu without
having to log out and scan the respective QR code. Fix logout logic in
AuthManager to properly update the waiter before switching to kitchen or
logging out.

Most importantly, something just clicked for me about how to use
Combine, so a lot of AuthManager was rewritten to take advantage of
proper Combine idioms for making publishers. A bit of that took place in
FirebaseExtensions too. Add a Future which lets you update the Firestore
document asynchronously.

Switch KitchenView to use the normal navigation hierarchy instead of a
weird, non-idiomatic Return button. A better title needs to be set if
the Back button looks weird.

Add a hack to AuthManager for remembering the kitchen title while the
hierarchy re-renders. Basically, if you do the switching from kitchen to
waiter mode via the debug menu, when the AuthManager.authState is
updated, SwiftUI will first re-render the KitchenView, by which point
AuthManager.kitchen property will be nil, and only then re-render
LaunchScreen when it's too late. In an ideal world, LaunchScreen would
be re-rendered immediately after changing .authState such that
KitchenView never even gets to the state where it tries to access
AuthManager.kitchen, but since I cba to force that somehow or write a
caching wrapper for KitchenView which would probably be more correct, we
just cache the kitchen name upon AuthManager and refresh it when needed.

There are some fixes to KitchenOrderOverView (which isn't titled
correctly btw but that's for a separate typo pass) such that it doesn't
refer to things via tons of unneeded hierarchy, but that was the thing
that I was gonna fix up later so that'll go in a future commit. Right
now we simply begin referring to the order directly instead of going
through KitchenWork (which is prone to deletion after refactoring just
like the QrScannerWork was.)

Sorry for the braindump and huge commit. It'd be too much effort to
decouple these changes to make it build at every point.