Models: Rework Firestore references
Note: This commit most definitely doesn't build and depends on the
following commits to fix things up.

I got kinda tired about the mess that was happening regarding the data
layer, so this commit is the first part of cleaning that up. The endgoal
is to unify the kitchen and waiter part data models so all of them refer
to the same objects, and storage logic can therefore be centralised.

Primary change that takes place here is both heavy usage of Combine to
manage asynchrony as well as the introduction of TypedDocumentReference
and TypedCollectionReference wrappers which provide nominal type safety
and convenience over raw Firestore references.

Most of the model attributes have been changed to read-only, given that
we consider the truth to lie with the server (i.e., Firestore) and no
local modifications should take place. Ideally all of the model
attributes should be immutable like that.

KitchenOrder is having its data model refactored. Hereby instead of
having a bunch of ExtraOrder objects in the Firestore that are wholly
decoupled from the Orders themselves, the parts are instead stored in an
array in a JSON serialized format. When needed, this array can be
appended to---Firestore can even do that atomically without replacing
the array itself.

Some document reference fields in Firestore have been made into proper
reference fields (notably the Table and User references upon Order) such
that it's easier to work with them instead of storing raw string IDs or
other such things. Again, ideally we would have all document references
look like that.

There's some fluff around aliasing Double to Currency (which is really
wrong btw! We should at least switch to using Decimal instead to prevent
accuracy loss, but I'll file that as an issue for another time) which is
currently temporary. In the future I hope to import the Currency type
that we previously had made for the older (standalone server) version.

This also includes a soft audit about referring to Firestore as
Firestore instead of Firebase, which is a different thing entirely. I
know, it's minutiae, but I like ontology to be followed well :)