AuthManager: Implement basic persistence of admin login state
This is, uh, let's just say "not great". There's no security and the
"admin" parameter is lost. The QR code scheme should be refactored in
gestalt, given that there's no real way to differentiate between
kitchens, but I'll put that in the backlog instead.

This ought to fix the linked issue but needs testing which I can't
perform right now.

See-Also: #2