Waiter: Fix active order detail view
This was a long-standing issue which was caused in part due to the new
data model. RestaurantOrder.OrderEntry used to (until this commit)
optionally hold a reference to its MenuItem, however this MenuItem
cannot be set easily when the order is loaded from the backend as
opposed to being assembled piecemeal via MenuView and its relatives.
This is exacerbated by all the model objects actually being structs and
generally not allowing modification.

Therefore the `item` reference had to go. This commit excises the `item`
attribute and replaces all its usages with explicit references to
`itemID` and having a so-called Menu (actually just a [MenuItem.FullID:
MenuItem] distionary) which is brought everywhere along with the order
itself so that the items can be looked up on demand.

This is not particularly pretty---I hope to someday come to implement an
OrderManager in the spirit of AuthManager which maintains the menu
internally instead of having it be passed around everywhere---but it
works and fixes the issue.

Another unprettiness is related to ActiveOrderList. It's possible for
orders to load before all the menu items have, in which case the order
list will be displayed, but the OrderDetailView cannot because it
assumes that MenuItems are readily available for things like calculating
totals. In that case, ActiveOrderListView will instead transition to a
spinner while waiting for the remaining items to load. (I initially
intended to solve this with a spinner overlay instead of a full
navigation, but eh, this works well enough too and is both simpler in
code and simpler to make work.)