Common+Waiter: Implement MenuManager
MenuManager is the preferred implementation for fetching menu contents
instead of doing it a-la-carte whenever it's needed, and therefore
having to have it in weird places. This is especially true when the menu
is needed both for placing orders and viewing active orders, but also
for the kitchen and stuff.

See-Also: 6d189ea, #7