Waiter: Fix ActiveOrderDetailView Add Items to Order
The bug that was there before meant that any changes to the
PartialOrderPart wouldn't be reflected in the DetailView. That's because
the object was kept in State but wasn't being observed, which it is now.

The simpler part of the fix is that we now always keep around a
PartialOrderPart and observe it, even before the user taps the button.
Whether its dependent contents are displayed (primarily the "Submit
Order" button) is determined by a separate variable.

However, the more troublesome part is that this runs into conflict
(again--see the TabViews within the waiter and kitchen interface) with
the way Swift initialisation works. To not overcomplicate things, the
exposed initialisation is done via the Wrapper view within, which takes
values from the environment and passes them on to the specific
initialiser.

I think I'll adopt this paradigm elsewhere (primarily within the root
views) sometime soon. It works out quite nicely, actually.