OrderManager: Don't replace orders being updating by us
Firestore's local component automatically calls the snapshot listener
even for changes that are initiated locally and which don't come from a
subscription. This has the downside that, in cases such as when adding
an extra part to an active order from the waiter flow, the extra part
becomes visible as persisted even before that's the case.

With this change, the document won't be replaced and taken from the
snapshot until it is for-sure persisted to the Firestore server and
fetched again, and will remain in the pre-update state until then.