Common: Factor out a reusable OrderDetailView
Given that both the waiter and the kitchen flow involves looking at the
order, this imho makes a lot of sense. Next up, removing the individual
implementations from both.

This also includes some debug convenience initializers for some models.
I intend to adopt them elsewhere so that IDs don't have to be written
down again and again (this ID desync in fact was a source of a bug that
I caught in the previews.)