Common/Models: Improve Firestore typesafe wrapping, alter Employee model
Sorry for amalgamating two different changes into a single commit again.

This started out as a refactoring of the Employee to use an `isManager`
field instead of the current `manager` to reflect their status. While I
was at it, I decided to merge the name/lastName into a single fullName
field to avoid ambiguities.

And while I was at _it_, I also decided to create a typesafe wrapper for
DocumentSnapshot so that we shouldn't have to use string keys
everywhere.

So this commit is encompassing all of that work. I added deprecations
for the old fields as well so that they can be gradually removed from
elsewhere in the codebase.