Waiter: Fix tab view crash when unwrapping optionals
Apparently there's literally no way to cache them, so instead we render
the empty view if the state ends up in a weird place.

This is basically the correct solution for the hack added in 4f0fc6c4
which did a similar thing to the kitchen view, given that SwiftUI
re-rendered the kitchen tab view first and LaunchScreen second, even
though that meant that the kitchen tab view got a nil for the kitchen
property access and that induced a crash.

Ideally if all values were passed through the environment, IfLet-s can
be used instead of hacking around it in the object portion of the code.