Waiter: Refactor PendingOrderPart, split up OrderDetailView
Remove an explicit MenuManager reference from PendingOrderPart. It was
only used for the subtotal, and it (similarly to RestaurantOrder) can be
calculated by passing in the menu on demand.

This in turn obviates most of the need for many Wrapper views, given
that initialization of a state object doesn't depend on an environment
object anymore. You can't imagine what a relief that is.

Relatedly, remove the massive monolith that was OrderDetailView and
instead expose its parts as reusable and isolated views which can be
combined on-demand to display order parts.

Propagate these changes to other places. Manually mark many animations
in ActiveOrderDetailView that were lost otherwise.